requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
python-dotenv
google-cloud-storage
fsspec
gcsfs
category_encoders
mlflow
memoized_property
psutil
termcolor
